app/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
prometheus-client==0.21.1
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
python-json-logger==3.2.1
psutil==6.1.1
httpx==0.28.1
pydantic==2.10.4
